Address review findings: move interfaces to consumer, add ErrNotFound
- Move SubscriberRepo/SettingsRepo interfaces to bot/repository.go as
  unexported types; delete internal/repository/interfaces.go. Bot no
  longer depends on the repository package for interface types.

- Add repository.ErrNotFound sentinel; GetGroupId now wraps
  mongo.ErrNoDocuments so callers are not coupled to the MongoDB driver.
  bot.Run() checks repository.ErrNotFound instead of mongo.ErrNoDocuments,
  and the mongo driver import is removed from bot/bot.go.
