+++ internal/repository/errors.go
@@ -1,0 +1,6 @@
+package repository
+
+import "errors"
+
+// ErrNotFound is returned when a requested document does not exist.
+var ErrNotFound = errors.New("not found")
